Rename adapters and requests to network interfaces and handlers
