+++ src/index.js
@@ -1,9 +1,9 @@
 import * as actions from './actions';
 import * as actionTypes from './constants/action-types';
 import * as httpMethods from './constants/http-methods';
+import * as networkInterfaces from './network-interfaces';
 import * as querySelectors from './selectors/query';
 
-export { default as superagentAdapter } from './adapters/superagent';
 export { default as connectRequest } from './components/connect-request';
 export { getQueryKey, reconcileQueryKey } from './lib/query-key';
 export { default as queriesReducer } from './reducers/queries';
@@ -11,4 +11,4 @@ export { default as entitiesReducer } from './reducers/entities';
 export { default as queryMiddleware } from './middleware/query';
 export { default as queryMiddlewareAdvanced } from './middleware/query-advanced';
 export { cancelQuery, mutateAsync, requestAsync, removeEntities, removeEntity } from './actions';
-export { actions, actionTypes, httpMethods, querySelectors };
+export { actions, actionTypes, httpMethods, networkInterfaces, querySelectors };
