+++ src/middleware/query.js
@@ -1,6 +1,6 @@
 import queryAdvanced from './query-advanced.js';
-import superagentAdapter from '../adapters/superagent';
+import superagentInterface from '../network-interfaces/superagent';
 
-const queryMiddleware = queryAdvanced(superagentAdapter);
+const queryMiddleware = queryAdvanced(superagentInterface);
 
 export default queryMiddleware;
