+++ src/network-interfaces/index.js
@@ -1,0 +1,1 @@
+export { default as superagent } from './superagent';
